Task-1/requirements.txt
streamlit==1.32.0
torch==2.2.0
diffusers==0.18.2
transformers==4.37.2
pillow==10.2.0
accelerate==0.27.2
huggingface_hub==0.19.4 
pymongo==4.11.0